move ncbi/common to ncbi/taxa_parsers
